+++ AddOns/AutoPotion/CHANGELOG.md
@@ -1,7 +1,13 @@
 # Auto Potion
 
-## [3.7.2](https://github.com/ollidiemaus/AutoPotion/tree/3.7.2) (2024-10-25)
-[Full Changelog](https://github.com/ollidiemaus/AutoPotion/compare/3.7.1...3.7.2) [Previous Releases](https://github.com/ollidiemaus/AutoPotion/releases)
+## [3.7.3](https://github.com/ollidiemaus/AutoPotion/tree/3.7.3) (2024-10-26)
+[Full Changelog](https://github.com/ollidiemaus/AutoPotion/compare/3.7.2...3.7.3) [Previous Releases](https://github.com/ollidiemaus/AutoPotion/releases)
 
 - Update AutoPotion.toc  
-- create default macro when mega macro is not in use (#57)  
+- Combat checks (#61)  
+    * add a stopcasting option  
+    * db option  
+    * update macroStr  
+    * tweak text  
+    * additional checks to ensure player is not in combat before EditMacro() is called  
+    * reduce wait to 0.5 seconds after player\_regen event | wrap editmacro() in pcall() to suppress LUA warnings